feat(forum): move views to www apps, override IndexView to prepare config.urls update
