+++ lacommunaute/www/forum/views.py
@@ -1,14 +1,57 @@
+import logging
+
+from django.conf import settings
 from django.contrib.auth.mixins import UserPassesTestMixin
-from django.db.models import Count
+from django.db.models import Count, Exists, OuterRef
 from django.shortcuts import get_object_or_404
+from django.urls import reverse
 from django.views.generic import DetailView, ListView
+from machina.apps.forum.views import ForumView as BaseForumView, IndexView as BaseIndexView
 from machina.core.loading import get_class
 
 from lacommunaute.forum.models import Forum
+from lacommunaute.forum_conversation.forms import PostForm
+from lacommunaute.forum_conversation.models import Topic
+from lacommunaute.users.models import User
 
+
+logger = logging.getLogger(__name__)
 
 ForumVisibilityContentTree = get_class("forum.visibility", "ForumVisibilityContentTree")
 PermissionRequiredMixin = get_class("forum_permission.viewmixins", "PermissionRequiredMixin")
+
+
+class IndexView(BaseIndexView):
+    pass
+
+
+class ForumView(BaseForumView):
+    paginate_by = settings.FORUM_TOPICS_NUMBER_PER_PAGE
+
+    def get_queryset(self):
+        forum = self.get_forum()
+        qs = (
+            forum.topics.exclude(type=Topic.TOPIC_ANNOUNCE)
+            .exclude(approved=False)
+            .annotate(likes=Count("likers"))
+            .annotate(has_liked=Exists(User.objects.filter(topic_likes=OuterRef("id"), id=self.request.user.id)))
+            .select_related(
+                "poster",
+                "poster__forum_profile",
+                "first_post",
+            )
+            .prefetch_related("poll", "poll__options", "poll__options__votes", "posts", "posts__attachments")
+            .order_by("-last_post_on")
+        )
+        return qs
+
+    def get_context_data(self, **kwargs):
+        forum = self.get_forum()
+        context = super().get_context_data(**kwargs)
+        context["FORUM_NUMBER_POSTS_PER_TOPIC"] = settings.FORUM_NUMBER_POSTS_PER_TOPIC
+        context["next_url"] = reverse("forum_extension:forum", kwargs={"pk": forum.pk, "slug": self.forum.slug})
+        context["form"] = PostForm(forum=forum, user=self.request.user)
+        return context
 
 
 class ModeratorEngagementView(PermissionRequiredMixin, ListView):
